tests/fixtures/fake_repo/requirements.txt
requests==2.25.0
flask==2.0.0
pyyaml==5.3
